Making some dacycler methods part of parent class for ease of maintenance
