+++ dabench/dacycler/_etkf.py
@@ -8,8 +8,7 @@ from jax.scipy import linalg
 import xarray as xr
 import xarray_jax as xj
 
-from dabench import dacycler, vector
-import dabench.dacycler._utils as dac_utils
+from dabench import dacycler
 
 
 class ETKF(dacycler.DACycler):
@@ -63,15 +62,6 @@ class ETKF(dacycler.DACycler):
                          ensemble=True,
                          B=B, R=R, H=H, h=h)
 
-    def _step_cycle(self, xb, obs_vals, obs_locs, obs_time_mask, obs_loc_mask,
-                    H=None, h=None, R=None, B=None):
-        if H is not None or h is None:
-            vals = self._cycle_obsop(
-                    xb, obs_vals, obs_locs, obs_time_mask,
-                    obs_loc_mask, H, R, B)
-            return vals
-        else:
-            return self._cycle_general_obsop(xb, yo, h, R, B)
 
     def _calc_default_H(self, obs_values, obs_loc_indices):
         H = jnp.zeros((obs_values.flatten().shape[0], self.system_dim))
@@ -108,7 +98,6 @@ class ETKF(dacycler.DACycler):
             else:
                 B = self.B
 
-        x0_xarray = x0_xarray.to_xarray()
         Xbt = x0_xarray[self._data_vars].to_array().data[0]
         nr,nc = Xbt.shape
         assert nr == self.ensemble_dim, (
@@ -116,8 +105,8 @@ class ETKF(dacycler.DACycler):
                     self.ensemble_dim, self.system_dim)
 
         # Apply obs masks to H
-        # H = jnp.where(obs_time_mask, H.T, 0).T
-        # H = jnp.where(obs_loc_mask.flatten(), H.T, 0).T
+        # H = jnp.where(obs_time_mask.flatten(), H.T, 0).T
+        H = jnp.where(obs_loc_mask.flatten(), H.T, 0).T
 
         # Analysis cycles over all obs in data_obs
         Xa = self._compute_analysis(Xb=Xbt.T,
@@ -222,9 +211,10 @@ class ETKF(dacycler.DACycler):
 
     def _cycle_and_forecast(self, cur_state, filtered_idx):
         # 1. Get data
-        # cur_state_vals = state[self.data_vars].data state_obs_tuple[0]
-
         # 1-b. Calculate obs_time_mask and restore filtered_idx to original values
+        cur_state = cur_state.to_xarray()
+        cur_time = cur_state['_cur_time'].data
+        cur_state = cur_state.drop_vars(['_cur_time'])
         obs_time_mask = filtered_idx > 0
         filtered_idx = filtered_idx - 1
 
@@ -241,99 +231,6 @@ class ETKF(dacycler.DACycler):
                 )
         # 3. Forecast next timestep
         next_state, forecast_states = self._step_forecast(analysis, n_steps=self.steps_per_window)
+        next_state = next_state.assign(_cur_time = cur_time + self.analysis_window)
 
         return xj.from_xarray(next_state), forecast_states
-
-    def cycle(self,
-              input_state,
-              start_time,
-              obs_vector,
-              n_cycles,
-              obs_error_sd=None,
-              analysis_window=0.2,
-              analysis_time_in_window=None,
-              return_forecast=False
-              ):
-        """Perform DA cycle repeatedly, including analysis and forecast
-
-        Args:
-            input_state (vector.StateVector): Input state.
-            start_time (float or datetime-like): Starting time.
-            obs_vector (vector.ObsVector): Observations vector.
-            n_cycles (int): Number of analysis cycles to run, each of length
-                analysis_window.
-            analysis_window (float): Time window from which to gather
-                observations for DA Cycle.
-            analysis_time_in_window (float): Where within analysis_window
-                to perform analysis. For example, 0.0 is the start of the
-                window. Default is None, which selects the middle of the
-                window.
-            return_forecast (bool): If True, returns forecast at each model
-                timestep. If False, returns only analyses, one per analysis
-                cycle. Default is False.
-
-        Returns:
-            vector.StateVector of analyses and times.
-        """
-
-        # These could be different if observer doesn't observe all variables
-        # For now, making them the same
-        self._observed_vars = obs_vector['variable'].values
-        self._data_vars = list(input_state.data_vars)
-
-        if obs_error_sd is None:
-            obs_error_sd = obs_vector.error_sd
-        self.analysis_window = analysis_window
-
-        # If don't specify analysis_time_in_window, is assumed to be middle
-        if analysis_time_in_window is None:
-            analysis_time_in_window = analysis_window/2
-
-        # Steps per window + 1 to include start
-        self.steps_per_window = round(analysis_window/self.delta_t) + 1
-
-        # Time offset from middle of time window, for gathering observations
-        _time_offset = (analysis_window/2) - analysis_time_in_window
-
-        # Set up for jax.lax.scan, which is very fast
-        all_times = dac_utils._get_all_times(
-            start_time,
-            analysis_window,
-            n_cycles)
-            
-
-        # Get the obs vectors for each analysis window
-        all_filtered_idx = dac_utils._get_obs_indices(
-            obs_times=jnp.array(obs_vector.time.values),
-            analysis_times=all_times+_time_offset,
-            start_inclusive=True,
-            end_inclusive=False,
-            analysis_window=analysis_window
-        )
-        
-        all_filtered_padded = dac_utils._pad_time_indices(all_filtered_idx, add_one=True)
-        self._obs_vector=obs_vector
-        self.obs_error_sd = obs_error_sd
-        if obs_vector.stationary_observers:
-            self._obs_loc_masks = jnp.ones(
-                obs_vector[self._observed_vars].to_array().shape, dtype=bool)
-        else:
-            self._obs_loc_masks = ~np.isnan(
-                obs_vector[self._observed_vars].to_array().data)[0]
-            self._obs_vector=self._obs_vector.fillna(0)
-        cur_state, all_values = jax.lax.scan(
-                self._cycle_and_forecast,
-                xj.from_xarray(input_state),
-                all_filtered_padded)
-                
-
-        all_vals_xr = xr.Dataset(
-            {var: (('cycle',) + tuple(all_values[var].dims),
-                   all_values[var].data)
-             for var in all_values.data_vars}
-        ).rename_dims({'time': 'cycle_timestep'})
-
-        if return_forecast:
-            return all_vals_xr
-        else:
-            return all_vals_xr.isel(cycle_timestep=0)
